move match_packet code generation to separate module
